chore: rename `scripts` to `utils`
