+++ packages/sw/src/sw.ts
@@ -6,10 +6,10 @@
 import { get } from 'idb-keyval';
 import * as Misskey from 'misskey-js';
 import type { PushNotificationDataMap } from '@/types.js';
-import type { I18n, Locale } from '@/scripts/i18n.js';
-import { createEmptyNotification, createNotification } from '@/scripts/create-notification.js';
-import { swLang } from '@/scripts/lang.js';
-import * as swos from '@/scripts/operations.js';
+import type { I18n, Locale } from '@/utils/i18n.js';
+import { createEmptyNotification, createNotification } from '@/utils/create-notification.js';
+import { swLang } from '@/utils/lang.js';
+import * as swos from '@/utils/operations.js';
 
 globalThis.addEventListener('install', () => {
 	// ev.waitUntil(globalThis.skipWaiting());
