+++ packages/sw/src/utils/lang.ts
@@ -7,7 +7,7 @@
  * Language manager for SW
  */
 import { get, set } from 'idb-keyval';
-import { I18n, type Locale } from '@/scripts/i18n.js';
+import { I18n, type Locale } from '@/utils/i18n.js';
 
 class SwLang {
 	public cacheName = `mk-cache-${_VERSION_}`;
