+++ packages/sw/src/utils/operations.ts
@@ -9,8 +9,8 @@
  */
 import * as Misskey from 'misskey-js';
 import type { SwMessage, SwMessageOrderType } from '@/types.js';
-import { getAccountFromId } from '@/scripts/get-account-from-id.js';
-import { getUrlWithLoginId } from '@/scripts/login-id.js';
+import { getAccountFromId } from '@/utils/get-account-from-id.js';
+import { getUrlWithLoginId } from '@/utils/login-id.js';
 
 export const cli = new Misskey.api.APIClient({ origin, fetch: (...args): Promise<Response> => fetch(...args) });
