+++ packages/sw/src/utils/twemoji-base.ts
